Eventos con frecuencia y notificacion de eventos

+++ src/main/java/utn/frba/dds/que_me_pongo/WebServices/Request/DeleteReservaRequest.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class DeleteReservaRequest {
     String uid;
-    Integer idEvento;
+    Long idEvento;
 }
